Move remaining passes to TheGraph.
