+++ src/apps/refactor_proto/refactor_proto/TheGraph.cpp
@@ -4,6 +4,7 @@
 #include "RendererReimplement.h" 
 
 #include "Camera.h"
+#include "DrawQuad.h"
 #include "Loader.h" // for semantics
 #include "Pass.h"
 #include "Profiling.h"
@@ -11,10 +12,54 @@
 #include "SetupDrawing.h"
 
 
+
 namespace ad::renderer {
 
 
 constexpr std::size_t gMaxDrawInstances = 2048;
+
+
+namespace {
+
+
+    /// @brief Layout compatible with the shader's `ViewBlock`
+    struct GpuViewBlock
+    {
+        GpuViewBlock(
+            math::AffineMatrix<4, GLfloat> aWorldToCamera,
+            math::Matrix<4, 4, GLfloat> aProjection 
+        ) :
+            mWorldToCamera{aWorldToCamera},
+            mProjection{aProjection},
+            mViewingProjection{aWorldToCamera * aProjection}
+        {}
+
+        GpuViewBlock(const Camera & aCamera) :
+            GpuViewBlock{aCamera.getParentToCamera(), aCamera.getProjection()}
+        {}
+
+        math::AffineMatrix<4, GLfloat> mWorldToCamera; 
+        math::Matrix<4, 4, GLfloat> mProjection; 
+        math::Matrix<4, 4, GLfloat> mViewingProjection;
+    };
+
+
+    void loadFrameUbo(const graphics::UniformBufferObject & aUbo)
+    {
+        GLfloat time =
+            std::chrono::duration_cast<std::chrono::milliseconds>(
+                std::chrono::high_resolution_clock::now().time_since_epoch()).count() / 1000.f;
+        proto::loadSingle(aUbo, time, graphics::BufferHint::DynamicDraw);
+    }
+
+
+    void loadCameraUbo(const graphics::UniformBufferObject & aUbo, const Camera & aCamera)
+    {
+        proto::loadSingle(aUbo, GpuViewBlock{aCamera}, graphics::BufferHint::DynamicDraw);
+    }
+
+
+} // unnamed namespace
 
 
 //
@@ -178,6 +223,67 @@ void TheGraph::loadDrawBuffers(const PartList & aPartList,
 }
 
 
+void TheGraph::renderFrame(const PartList & aPartList, const Camera & aCamera, Storage & aStorage)
+{
+    {
+        PROFILER_SCOPE_RECURRING_SECTION("load_frame_UBOs", CpuTime, GpuTime, BufferMemoryWritten);
+        loadFrameUbo(*mUbos.mFrameUbo);
+        // Note in a more realistic application, several cameras would be used per frame.
+        loadCameraUbo(*mUbos.mViewingUbo, aCamera);
+    }
+
+    // Use the same indirect buffer for all drawings
+    // Its content will be rewritten by distinct passes though
+    // With the current approach, this binding could be done only once at construction
+    graphics::bind(mIndirectBuffer, graphics::BufferType::DrawIndirect);
+
+    passDepth(aPartList, aStorage);
+    passForward(aPartList, aStorage);
+
+    //
+    // Debug rendering of the depth texture
+    //
+    auto [nearZ, farZ] = getNearFarPlanes(aCamera);
+    showDepthTexture(mDepthMap, nearZ, farZ) ;
+}
+
+
+void TheGraph::showDepthTexture(const graphics::Texture & aTexture,
+                                float aNearZ, float aFarZ,
+                                unsigned int aStackPosition)
+{
+    assert(aStackPosition < 4);
+
+    // Note: with stencil, we could draw those rectangles first,
+    // and prevent main rasterization behind them.
+
+    glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0);
+    glActiveTexture(GL_TEXTURE0);
+    glBindTexture(aTexture.mTarget, aTexture);
+
+    // Would require scissor test to clear only part of the screen.
+    //gl.Clear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
+
+    {
+        glDisable(GL_DEPTH_TEST);
+        glDisable(GL_BLEND);
+        glDepthMask(GL_FALSE);
+    }
+
+    math::Size<2, int> fbSize = mGlfwAppInterface->getFramebufferSize();
+
+    glViewport(0,
+               aStackPosition * fbSize.height() / 4,
+               fbSize.width() / 4,
+               fbSize.height() / 4);
+
+    drawQuad({
+        .mSourceChannel = 0,
+        .mLinearization = DrawQuadParameters::DepthMethod1,
+        .mNearDistance = aNearZ,
+        .mFarDistance = aFarZ,
+    });
+}
 
 void TheGraph::passDepth(const PartList & aPartList, Storage & aStorage)
 {
@@ -241,5 +347,54 @@ void TheGraph::passDepth(const PartList & aPartList, Storage & aStorage)
     }
 }
 
+
+void TheGraph::passForward(const PartList & aPartList, Storage & mStorage)
+{
+    PROFILER_SCOPE_RECURRING_SECTION("pass_forward", CpuTime, GpuTime);
+
+    // Can be done once for distinct camera, if there is no culling
+    PassCache passCache = preparePass("forward", aPartList, mStorage);
+
+    // Load the data for the part and pass related UBOs (TODO: SSBOs)
+    loadDrawBuffers(aPartList, passCache);
+
+    // FBO (default framebuffer)
+    glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0);
+
+    //
+    // Set pipeline state
+    //
+    {
+        PROFILER_SCOPE_RECURRING_SECTION("set_pipeline_state", CpuTime);
+        // TODO handle pipeline state with an abstraction
+        glEnable(GL_CULL_FACE);
+        glEnable(GL_DEPTH_TEST);
+        glDepthMask(GL_TRUE);
+
+        // We implemented alpha testing (cut-out), no blending.
+        glDisable(GL_BLEND);
+    }
+
+    gl.Clear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
+
+    // Might loop over cameras, or any other variation
+    //for(whatever dimension)
+    {
+        // TODO should be done once per viewport
+        glViewport(0, 0,
+                    mGlfwAppInterface->getFramebufferSize().width(),
+                    mGlfwAppInterface->getFramebufferSize().height());
+
+        // Reset texture bindings (to make sure that texturing does not work by "accident")
+        // this could be extended to reset everything in the context.
+        glBindTexture(GL_TEXTURE_2D, 0);
+        glBindTexture(GL_TEXTURE_2D_ARRAY, 0);
+
+        {
+            PROFILER_SCOPE_RECURRING_SECTION("draw_instances", CpuTime, GpuTime, GpuPrimitiveGen, DrawCalls, BufferMemoryWritten);
+            draw(passCache, mUbos.mUboRepository, mDummyTextureRepository);
+        }
+    }
+}
 
 } // namespace ad::renderer
